:sparkles: (nordigen): ability to set secrets via the UI
