+++ packages/desktop-client/src/components/modals/NordigenExternalMsg.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState, useRef } from 'react';
 
 import { send } from 'loot-core/src/platform/client/fetch';
 
+import useNordigenStatus from '../../hooks/useNordigenStatus';
 import AnimatedLoading from '../../icons/AnimatedLoading';
 import { colors } from '../../style';
 import { Error, Warning } from '../alerts';
@@ -36,29 +37,6 @@ function useAvailableBanks(country) {
 
   return {
     data: banks,
-    isLoading,
-  };
-}
-
-function useNordigenStatus() {
-  const [configured, setConfigured] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-
-  useEffect(() => {
-    async function fetch() {
-      setIsLoading(true);
-
-      const results = await send('nordigen-status');
-
-      setConfigured(results.configured || false);
-      setIsLoading(false);
-    }
-
-    fetch();
-  }, [setConfigured, setIsLoading]);
-
-  return {
-    configured,
     isLoading,
   };
 }
